+++ tests/Spec/GenericSpec.hs
@@ -12,7 +12,8 @@ module Spec.GenericSpec (main, spec) where
 
 import Data.Proxy (Proxy(..))
 import Instances.Generic ()
-import Spec.Utils (matchesTextShowSpec, genericTextShowSpec)
+import Instances.Utils (GenericExample)
+import Spec.Utils (matchesTextShowSpec, matchesTextShow1Spec, genericTextShowSpec)
 import Test.Hspec (Spec, describe, hspec, parallel)
 import TextShow.Generic (ConType)
 
@@ -20,8 +21,14 @@ main :: IO ()
 main = hspec spec
 
 spec :: Spec
-spec = parallel . describe "ConType" $ do
-    let p :: Proxy ConType
-        p = Proxy
-    matchesTextShowSpec p
-    genericTextShowSpec p
+spec = parallel $ do
+    describe "ConType" $ do
+        let p :: Proxy ConType
+            p = Proxy
+        matchesTextShowSpec p
+        genericTextShowSpec p
+    describe "GenericExample Int" $ do
+        let p :: Proxy (GenericExample Int)
+            p = Proxy
+        matchesTextShowSpec  p
+        matchesTextShow1Spec p